+++ lib/filename_cleaner.rb
@@ -1,4 +1,4 @@
-require_relative './filename_cleaner/version'
-require_relative './filename_cleaner/cli'
-require_relative './filename_cleaner/filename_cleaner'
-require_relative './filename_cleaner/core_ext/object/blank'
+require_relative "./filename_cleaner/version"
+require_relative "./filename_cleaner/cli"
+require_relative "./filename_cleaner/filename_cleaner"
+require_relative "./filename_cleaner/core_ext/object/blank"