+++ lib/filename_cleaner/filename_cleaner.rb
@@ -1,12 +1,12 @@
 module FilenameCleaner
-  DOT = '.'
+  DOT = "."
   class << self
     # Sanitize the any name with or without any extension
     #
     # @param [String] name the input string
     # @param [String] sep_char the separator char to be used
     # @return [String] output string with special chars replaced withe specified string
-    def sanitize(name, sep_char = '.', have_extension = false)
+    def sanitize(name, sep_char = ".", have_extension = false)
       if have_extension
         sanitize_name_with_extension(name, sep_char)
       else
@@ -14,14 +14,14 @@ module FilenameCleaner
       end
     end
 
-    private
+  private
 
     # Sanitize the any name with or without any extension
     #
     # @param [String] name the input string
     # @param [String] sep_char the separator char to be used
     # @return [String] output string with special chars replaced withe specified string
-    def sanitize_name_without_extension(name, sep_char = '.')
+    def sanitize_name_without_extension(name, sep_char = ".")
       replace_dot(sanitize_with_dot(name), sep_char)
     end
 
@@ -29,7 +29,7 @@ module FilenameCleaner
     #
     # @param [String] name the input filename with extension
     # @return [String] the output file with special characters replaced
-    def sanitize_name_with_extension(name, sep_char = '.')
+    def sanitize_name_with_extension(name, sep_char = ".")
       extension = File.extname(name)
       name_only = File.basename(name, ".*")
       name_only = replace_dot(sanitize_with_dot(name_only), sep_char)
@@ -45,7 +45,7 @@ module FilenameCleaner
       name.gsub!(/[^0-9A-Za-z\-_ ]/, DOT)
 
       # Replace multiple occurrences of a given character with a dot
-      ['-', '_', ' '].each do |c|
+      ["-", "_", " "].each do |c|
         name.gsub!(/#{Regexp.quote(c)}+/, DOT)
       end
 
@@ -53,7 +53,7 @@ module FilenameCleaner
       name.gsub!(/#{Regexp.quote(DOT)}+/, DOT)
 
       # Remove the last char if it is a dot
-      name.gsub!(/\.$/, '') if name[-1] == DOT
+      name.gsub!(/\.$/, "") if name[-1] == DOT
 
       # return the result
       name