+++ test/test_helper.rb
@@ -1,10 +1,10 @@
-require 'minitest'
-require 'minitest/autorun'
-require 'minitest/pride'
-require 'minitest-spec-context'
-require 'coveralls'
-require 'pry'
-require 'awesome_print'
-require_relative '../lib/filename_cleaner'
-include FilenameCleaner
+require "coveralls"
 Coveralls.wear!
+require "minitest"
+require "minitest/autorun"
+require "minitest/pride"
+require "minitest-spec-context"
+require "pry"
+require "awesome_print"
+require_relative "../lib/filename_cleaner"
+include FilenameCleaner